Rename lib/helpers/pivotaly to lib/helpers/state
