+++ lib/commands/deliverStory.js
@@ -1,5 +1,5 @@
 const PtStory = require('../../model/stories')
-const {getState} = require('../helpers/pivotaly')
+const {getState} = require('../helpers/state')
 const rebounds = require('../validation/rebounds')
 
 exports.deliverStory = async context => {
