+++ lib/commands/deliverTask.js
@@ -1,6 +1,6 @@
 const PtTask = require('../../model/tasks')
 const rebounds = require('../validation/rebounds')
-const {getState} = require('../helpers/pivotaly')
+const {getState} = require('../helpers/state')
 
 module.exports = async (taskTreeItem, context) => {
   if(!taskTreeItem.hasOwnProperty('itemId')) return rebounds('failedDeliveredTask')
