+++ lib/commands/finishStory.js
@@ -1,6 +1,6 @@
 const PtStory = require('../../model/stories')
 const rebounds = require('../validation/rebounds')
-const {getState} = require('../helpers/pivotaly')
+const {getState} = require('../helpers/state')
 
 exports.finishStory = async context => {
   const state = getState(context)
