+++ lib/commands/finishStory.spec.js
@@ -3,7 +3,7 @@ jest.mock("../validation/rebounds")
 jest.mock("../../model/stories")
 const PtStory = require("../../model/stories")
 const rebounds = require("../validation/rebounds")
-const {getState} = require("../helpers/pivotaly")
+const {getState} = require("../helpers/state")
 const { finishStory } = require('./finishStory')
 const Context = require('../../test/factories/vscode/context')
