+++ lib/commands/linkStory.js
@@ -1,5 +1,5 @@
 const {window} = require('vscode')
-const {setState} = require('../helpers/pivotaly')
+const {setState} = require('../helpers/state')
 const {getActiveBranch} = require('../helpers/git')
 const {validate} = require('../validation/validate')
 const rebounds = require('../validation/rebounds')
