+++ lib/commands/refreshStateView.js
@@ -1,4 +1,4 @@
-const {refreshState} = require('../helpers/pivotaly')
+const {refreshState} = require('../helpers/state')
 
 module.exports = async (context, dataProvider) => {
   await refreshState(context)
