+++ lib/commands/refreshStateView.spec.js
@@ -2,7 +2,7 @@ jest.mock('../helpers/pivotaly')
 const refreshStateView = require('./refreshStateView')
 const StateInfoDataProvider = require('../views/storyInfoDataProvider')
 const Context = require('../../test/factories/vscode/context')
-const ptlyHelpers = require('../helpers/pivotaly')
+const ptlyHelpers = require('../helpers/state')
 
 const context = Context.build()
