+++ lib/commands/showAllCommands.js
@@ -1,6 +1,6 @@
 const {window, commands} = require("vscode")
 const _ = require('lodash')
-const {getState} = require("../helpers/pivotaly")
+const {getState} = require("../helpers/state")
 const {storyState, workState, statistics} = require("./commands").commands
 const getAllKeys = require('../adapters/getAllKeys')
 const {model} = require('../../model')
