+++ lib/helpers/state.spec.js
@@ -1,21 +1,21 @@
 jest.mock('../../model/stories')
-const pivotaly = require('./pivotaly')
+const state = require('./state')
 const Context = require('../../test/factories/vscode/context')
 const PtStory = require('../../model/stories')
 
 const context = Context.build()
 
-describe('#pivotaly.getState', () => {
+describe('#state.getState', () => {
   it('returns state', async () => {
     // ensure .get(state) was called
-    const result = await pivotaly.getState(context)
+    const result = await state.getState(context)
     expect(result).toHaveProperty('branch')
     expect(result).toHaveProperty('story')
     expect(result).toHaveProperty('isChore')
   })
 })
 
-describe('#pivotaly.setState', () => {
+describe('#state.setState', () => {
   beforeAll(() => {
     PtStory.mockImplementation(() => ({
       getStory: () => ({data: {story_type: 'chore'}})
@@ -25,13 +25,13 @@ describe('#pivotaly.setState', () => {
   afterAll(() => jest.clearAllMocks())
 
   it('it should update set chore if branch and story are not provided', async () => {
-    const result = await pivotaly.setState(context)
+    const result = await state.setState(context)
     const newState = JSON.parse(result)
     expect(newState).toHaveProperty('isChore', true)
   })
 
   it('should update all props given branch and story', async () => {
-    const result = await pivotaly.setState(context, 'branch3', '125')
+    const result = await state.setState(context, 'branch3', '125')
     const newState = JSON.parse(result)
     expect(newState).toHaveProperty('isChore', true)
     expect(newState).toHaveProperty('branch', 'branch3')
@@ -40,9 +40,9 @@ describe('#pivotaly.setState', () => {
   })
 })
 
-describe('#pivotaly.wipeState', () => {
+describe('#state.wipeState', () => {
   it('should set all properties to undefined', async () => {
-    const newState = await pivotaly.wipeState(context)
+    const newState = await state.wipeState(context)
     expect(newState).toBe('{}')
   })
 })
