+++ lib/validation/validate.js
@@ -3,7 +3,7 @@ const {getValidationChain} = require('./validationChains')
 const rebounds = require('./rebounds')
 const {getActiveBranch} = require("../helpers/git")
 const {didBranchChange} = require("./validators/didBranchChange")
-const {setState} = require("../helpers/pivotaly")
+const {setState} = require("../helpers/state")
 const {getStoryID} = require("../helpers/stories")
 const {isBranchDefault} = require("./validators/isBranchDefault")
