+++ lib/validation/validators/didBranchChange.js
@@ -1,5 +1,5 @@
 const {getActiveBranch} = require("../../helpers/git")
-const {getState} = require("../../helpers/pivotaly")
+const {getState} = require("../../helpers/state")
 
 exports.didBranchChange = async (context) => {
   const pivotalyState = getState(context)
