+++ lib/validation/validators/isStoryIDValid.js
@@ -1,5 +1,5 @@
 const PtStory = require('../../../model/stories')
-const {getState} = require('../../helpers/pivotaly')
+const {getState} = require('../../helpers/state')
 
 exports.isStoryIDValid = async (context, storyId = null) => {
   const validStoryId = storyId || getState(context).story
