+++ lib/views/storyInfoDataProvider.js
@@ -1,5 +1,5 @@
 const {TreeItem, TreeItemCollapsibleState, EventEmitter} = require('vscode')
-const {getState} = require('../helpers/pivotaly')
+const {getState} = require('../helpers/state')
 const PtTask = require('../../model/tasks')
 const PtBlocker = require('../../model/blockers')
 const TaskAndBlockerTreeItem = require('./taskAndBlockerTreeItem')
